refactor(frontend): import build-hash handshake modules directly
