--- packages/frontend/src/features/buildHashHandshake/index.ts
@@ -1,2 +1,0 @@
-export { default as BuildSkewRefresher } from './BuildSkewRefresher';
-export { recordServerBuildHash } from './buildHashHandshake';
